Refactor: JoinRequest 정규식 , PasswordEncrypt 패키지이동
1. 기존에 작성했던 이메일정규식이 정상 작동하지않아 JoinRequest 이메일정규식을 변경했다.

2. PasswordEncrypt는 user말고 다른곳에서 쓰일거라는 생각이 들지않아 공통 util클래스에서 user패키지 안에 이동했다. 이부분에 대해선 이야기를 나눠봐야할것같다.
